requirements.txt
beautifulsoup4==4.9.3
celery==5.0.5
django-celery-results==2.0.1
django-celery-beat==2.2.0
django-ckeditor==6.1.0
django-environ==0.4.5
django-filter==2.4.0
django-rest-auth==0.9.5
django-web3-auth==0.1.2
django-split-settings==1.0.1
Django==3.2.2
djangorestframework==3.12.4
drf-yasg==1.20.0
ethereum==2.3.2
Pillow==8.2.0
psycopg2==2.8.6
python-memcached==1.59
pyppeteer==0.2.5
requests==2.25.1
rlp==1.2.0
sentry-sdk==1.1.0
web3==5.19.0
